Refactor attempt and answer endpoints.
